Implement project function
Some cifti metric files do not store data for every single vertex/voxel
of the structures the represent. For example, a cortical thickness file
may not have data for the intrahemispheric, non-surface portion of the
mesh. This makes it difficult to use and visualize surfaces in python,
as the matrix will be a different size than the mesh matrix. The project
function works after the cifti image has been filtered down to voxels or
vertices (voxels doesn't work yet though) and a single structure. It
projects the data into a matrix of expected size for the structure, as
recorded in the cifti file metadata.